+++ Code/Modules/MeanShift/otbMeanShiftModuleController.h
@@ -70,7 +70,7 @@ public:
   virtual void SwitchClusters(bool sc);
   virtual void SwitchBoundaries(bool sb);
   virtual void SetSpatialRadius(unsigned int sr);
-  virtual void SetSpectralRadius(unsigned int sr);
+  virtual void SetSpectralRadius(double sr);
   virtual void SetMinRegionSize(unsigned int mr);
   virtual void SetOpacity(double op);
   virtual void UpdateViewerDisplay();
